minor refactoring of tests
