+++ pypeman/tests/test_msgstore.py
@@ -8,17 +8,9 @@ from pypeman import msgstore
 from pypeman import nodes
 from pypeman.channels import BaseChannel
 from pypeman.test import TearDownProjectTestCase as TestCase
-from pypeman.tests.common import TestException, generate_msg
-
-
-class TestNode(nodes.BaseNode):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Used to test if node is processed during test
-
-    def process(self, msg):
-        print("Process %s" % self.name)
-        return msg
+from pypeman.tests.common import generate_msg
+from pypeman.tests.common import TestException
+from pypeman.tests.common import TestNode
 
 
 class TestConditionalErrorNode(nodes.BaseNode):
@@ -30,13 +22,6 @@ class TestConditionalErrorNode(nodes.BaseNode):
             raise TestException()
 
         return msg
-
-
-class ExceptNode(TestNode):
-    # This node raises an exception
-    def process(self, msg):
-        super().process(msg)
-        raise TestException()
 
 
 class MsgstoreTests(TestCase):
